Split TypeId into separate file
